moving astar code to path.js
